Add methods to handle Unrevsolved Package Message
