+++ consumer.py
@@ -24,9 +24,13 @@ import os
 
 from thoth.investigator import __service_version__
 
-from thoth.messaging import MessageBase, UnresolvedPackageMessage, SolvedPackageMessage
+from thoth.messaging import MessageBase
+from thoth.messaging import UnresolvedPackageMessage
+from thoth.messaging import UnrevsolvedPackageMessage
+from thoth.messaging import SolvedPackageMessage
 from thoth.investigator.investigate_unresolved_package import parse_unresolved_package_message
 from thoth.investigator.investigate_solved_package import parse_solved_package_message
+from thoth.investigator.investigate_unrevsolved_package import parse_revsolved_package_message
 
 from thoth.common import OpenShift
 from thoth.storages.graph import GraphDatabase
@@ -48,6 +52,7 @@ _LOGGER.info("Thoth Investigator consumer v%s", __service_version__)
 # initialize the application
 app = MessageBase().app
 unresolved_package_message_topic = UnresolvedPackageMessage().topic
+unrevsolved_package_message_topic = UnrevsolvedPackageMessage().topic
 solved_package_message_topic = SolvedPackageMessage().topic
 
 openshift = OpenShift()
@@ -74,6 +79,13 @@ async def consume_unresolved_package(unresolved_packages) -> None:
     """Loop when an unresolved package message is received."""
     async for unresolved_package in unresolved_packages:
         parse_unresolved_package_message(unresolved_package=unresolved_package, openshift=openshift, graph=graph)
+
+
+@app.agent(unrevsolved_package_message_topic)
+async def consume_unrevsolved_package(unrevsolved_packages) -> None:
+    """Loop when an unresolved package message is received."""
+    async for unrevsolved_package in unrevsolved_packages:
+        parse_revsolved_package_message(unrevsolved_package=unrevsolved_package, openshift=openshift)
 
 
 @app.agent(solved_package_message_topic)
